Removed duplicated code and applied open-closed principle.
