+++ AirCasting/SearchAndFollow/CompleteScreen/Views/SearchAndFollowChartViewModel.swift
@@ -4,50 +4,27 @@
 import Foundation
 import SwiftUI
 
+struct ChartMeasurement {
+    let value: Double
+    let time: Date
+}
+
 class SearchAndFollowChartViewModel: ObservableObject {
     struct ChartDot {
         let value: Double
         let color: Color
     }
     
-    struct ChartMeasurement {
-        let value: Double
-        let time: Date
-    }
-    
-    enum EntriesSensor: String {
-        case OpenAQ = "OpenAQ"
-        case AirBeam = "AirBeam"
-        case PurpleAir = "PurpleAir"
-        case undefined = ""
-    }
-    
     @Published var entries: [ChartDot] = []
     
     let numberOfEntries = 9
     
-    func generateEntries(with measurements: [ChartMeasurement], thresholds: ThresholdsValue, basedOn sensorName: String) -> (Date?, Date?) {
+    func generateEntries(with measurements: [ChartMeasurement], thresholds: ThresholdsValue, basedOn sensor: ChartSensor) -> (Date?, Date?) {
         var times: [Date] = []
         var buffer: [ChartMeasurement] = []
         var updatedMeasurements = measurements
         
-        let sensor = getEntriesSensor(using: sensorName)
-        if sensor != .OpenAQ {
-            guard let firstElement = updatedMeasurements.reversed().first?.time else { return (nil, nil) }
-            
-            switch sensor {
-            case .AirBeam:
-                // AB stands for AirBeam
-                clearABData(using: &updatedMeasurements, hourToRemove: firstElement)
-            case .PurpleAir:
-                // PA stands for PurpleAir
-                clearPAData(using: &updatedMeasurements, hourToRemove: firstElement)
-            case .undefined:
-                Log.info("Missing sensor name in the chart VM")
-            default:
-                return (nil, nil)
-            }
-        }
+        sensor.clean(measurements: &updatedMeasurements)
         
         for measurement in updatedMeasurements.reversed() {
             if entries.count == 9 {
@@ -71,35 +48,6 @@ class SearchAndFollowChartViewModel: ObservableObject {
         entries.reverse()
         return (startTime: times.min(), endTime: times.max())
     }
-    
-    private func getEntriesSensor(using sensor: String) -> EntriesSensor {
-        switch sensor {
-        case EntriesSensor.OpenAQ.rawValue: return .OpenAQ
-        case EntriesSensor.AirBeam.rawValue: return .AirBeam
-        case EntriesSensor.PurpleAir.rawValue: return .PurpleAir
-        default: return .undefined
-        }
-    }
-    
-    private func clearPAData(using updatedMeasurements: inout [ChartMeasurement], hourToRemove: Date) {
-        for (index, item) in updatedMeasurements.enumerated().reversed() {
-            if Calendar.current.component(.hour, from: item.time) == hoursForRemoval(using: hourToRemove) && Calendar.current.component(.minute, from: item.time) != 00 {
-                updatedMeasurements.remove(at: index)
-            }
-        }
-    }
-    
-    private func clearABData(using updatedMeasurements: inout [ChartMeasurement], hourToRemove: Date) {
-        for (index, item) in updatedMeasurements.enumerated().reversed() {
-            if Calendar.current.component(.hour, from: item.time) == hoursForRemoval(using: hourToRemove) {
-                updatedMeasurements.remove(at: index)
-                continue
-            }
-            break
-        }
-    }
-    
-    private func hoursForRemoval(using element: Date) -> Int { Calendar.current.component(.hour, from: element) }
     
     private func addAverage(for buffer: [ChartMeasurement], times: inout [Date], thresholds: ThresholdsValue) {
         guard !buffer.isEmpty else { return }
